feat : la possibilter de naviger simplement entre dashboard et home page

+++ src/lib/frontend/redux/features/loginSlice.js
@@ -13,16 +13,18 @@ export const login = createAsyncThunk('auth/login', async (data, { rejectWithVal
     try {
         const response = await axios.post(`${process.env.NEXT_PUBLIC_BACKEND_CLIENT_URL}/api/auth/login`, data);
         const { token, existingUser } = response.data;
-        Cookies.set('token', token, {
-            httpOnly: false, 
-            secure: process.env.NODE_ENV,
-            sameSite: 'strict', 
-            expires: 1, 
-          });
-        localStorage.setItem('token', token);
-        localStorage.setItem('fullName', existingUser.fullName);
-        localStorage.setItem('email', existingUser.email);
-        localStorage.setItem('id', existingUser._id);
+        if(token && existingUser){
+            Cookies.set('token', token, {
+                httpOnly: false, 
+                secure: process.env.NODE_ENV,
+                sameSite: 'strict', 
+                expires: 1, 
+              });
+            localStorage.setItem('token', token);
+            localStorage.setItem('fullName', existingUser.fullName);
+            localStorage.setItem('email', existingUser.email);
+            localStorage.setItem('id', existingUser._id);
+        }
         return response.data
     } catch (error) {
         return rejectWithValue(error.response?.data || error.message);
